+++ lib/bones_api_server.dart
@@ -3,5 +3,6 @@ library;
 
 export 'bones_api.dart';
 export 'src/bones_api_error_zone.dart';
+export 'src/bones_api_extension_io.dart';
 export 'src/bones_api_hotreload.dart';
 export 'src/bones_api_server.dart';